将type选项改为mode
