+++ src/Mode/IMode.php
@@ -7,12 +7,25 @@ namespace CurdGen\Mode;
 interface IMode
 {
 
+    public function modelValidate($columns_set);
+    public function modelAuto($columns_set);
+
     public function tableColumn($columns_set);
     public function formColumn($columns_set);
     public function formExtra($columns_set);
     public function editColumn($columns_set);
     public function saveColumn($columns_set);
-    public function addTopButton(&$stub, &$add_btn_placeholder = null);
-    public function editRightButton(&$stub, &$table_btn_placeholder = null);
+
+    public function injectAddTopButton();
+    public function injectEditRightButton();
+    public function injectSaveTopButton();
+
+    public function funSuccessJumpUrl();
+    public function funFormDisplay();
+    public function editMetaTitle();
+    public function addMetaTitle();
+    public function editExtraFun();
+    public function addExtraFun();
+    public function tableBtnPlaceholder();
 
 }
