--- src/Mode/Type/Float/Operate/IOperate.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace CurdGen\Mode\Type\Float\Operate;
-
-interface IOperate
-{
-
-    public function modalFunParse();
-
-    public function tableBtnPlaceholderParse();
-
-}
