--- src/Operate/IOperate.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace CurdGen\Operate;
-
-
-interface IOperate
-{
-
-    public function operateParse();
-
-}
