--- src/Operate/IRightButton.php
@@ -1,11 +1,0 @@
-<?php
-
-
-namespace CurdGen\Operate;
-
-
-interface IRightButton
-{
-    public function rightButtonParse();
-
-}
