--- src/Operate/ITopButton.php
@@ -1,11 +1,0 @@
-<?php
-
-
-namespace CurdGen\Operate;
-
-
-interface ITopButton
-{
-    public function topButtonParse();
-
-}
