+++ src/Stubs/controller.stub
@@ -45,7 +45,55 @@ class {DummyModel}Controller extends QsListController{
 {DummySave}
 
 
-{DummyEdit}
+    {DummyEditExtraFun}
+
+    protected function addBuilderFormItem(&$builder){
+        {DummyFormExtra}
+
+        $builder
+        {DummyFormColumns}
+        ;
+    }
+
+    public function edit($id=null){
+        if (IS_POST) {
+            parent::autoCheckToken();
+            $m_id = I('post.id');
+            $data = I('post.');
+            $model = D('{DummyModel}');
+            if(!$m_id){
+                E('缺少{DummyModelTitle}ID');
+            }
+
+            $ent = $model->getOne($m_id);
+            if(!$ent){
+                E('不存在{DummyModelTitle}');
+            }
+
+            {DummyEditColumns}
+            if($model->createSave($ent) === false){
+                $this->error($model->getError());
+            }
+            else{
+                sysLogs('修改{DummyModelTitle}, id:' . $m_id);
+                $this->success('修改成功', '{DummyFunSuccessJumpUrl}');
+            }
+        } else {
+            $info = D('{DummyModel}')->getOne($id);
+
+            $builder = new \Qscmf\Builder\FormBuilder();
+
+            $this->addBuilderFormItem($builder);
+
+            $builder->setMetaTitle('{DummyEditMetaTitle}')
+                ->setPostUrl(U('edit'))
+                ->addFormItem('id', 'hidden', 'ID')
+                ->setFormData($info)
+                {DummyFunFormDisplay}
+
+
+        }
+    }
 
 
     public function forbid(){
@@ -95,6 +143,43 @@ class {DummyModel}Controller extends QsListController{
     }
 
 
-{DummyAdd}
+    {DummyAddExtraFun}
+
+
+    public function add(){
+        if (IS_POST) {
+            parent::autoCheckToken();
+            $data = I('post.');
+
+            $model = D('{DummyModel}');
+            $r = $model->createAdd($data);
+            if($r === false){
+                $this->error($model->getError());
+            }
+            else{
+                sysLogs('新增{DummyModelTitle}, id:' . $r);
+
+
+                $this->success(l('add') . l('success'), '{DummyFunSuccessJumpUrl}');
+            }
+        }
+        else {
+            $builder = new \Qscmf\Builder\FormBuilder();
+
+            $data_list = array(
+                "status"=>1
+            );
+
+            if($data_list){
+                $builder->setFormData($data_list);
+            }
+
+            $this->addBuilderFormItem($builder);
+
+            $builder->setMetaTitle('{DummyAddMetaTitle}')
+                ->setPostUrl(U('add'))
+                {DummyFunFormDisplay}
+        }
+    }
 
 }
